bot/requirements.txt
# Telegram Bot Dependencies
aiogram==3.4.1
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1
loguru==0.7.2
python-dotenv==1.0.0
aiohttp>=3.9.0  # For Yandex Vision API

# OCR Dependencies
pytesseract==0.3.10
Pillow==10.2.0
